Extend method comment and use size_t for num_stripes.

+++ cuckoo_index.h
@@ -30,7 +30,7 @@ class CuckooIndex : public IndexStructure {
  public:
   bool StripeContains(size_t stripe_id, int value) const override;
 
-  Bitmap64 GetQualifyingStripes(int value, int num_stripes) const override;
+  Bitmap64 GetQualifyingStripes(int value, size_t num_stripes) const override;
 
   std::string name() const override { return name_; }
